New start.sh & restart.sh + new Dockerfile

+++ server.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 while true; do
-	kill -9 $(ps -a | grep nc | awk '{print $1}')
+	if ps -a | grep nc > /dev/null; then
+	    kill -9 $(ps -a | grep nc | awk '{print $1}')
+	fi
 	printf "HTTP/1.0 200 OK\nContent-Length: `wc index.html | awk '{print $3-1}'`\n\n`cat index.html`" | nc -l 9191 > /dev/null
 done;
